requirements.txt
autopep8==1.6.0
blinker==1.4
click==8.0.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Mail==0.9.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycodestyle==2.8.0
python-dotenv==0.19.2
toml==0.10.2
Werkzeug==2.0.3
WTForms==3.0.1
